Fix double

+++ nff/data/dataset.py
@@ -376,7 +376,7 @@ def convert_nan(x):
             if type(y) in [int, float]:
                 new_x.append(torch.Tensor(y))
             elif isinstance(y, torch.Tensor):
-                new_x.append(y.double())
+                new_x.append(y.float())
         else:
             new_x.append(y)
